Use TFLite inference for stable Render deployment

+++ utils/preprocess.py
@@ -12,6 +12,7 @@ def preprocess_image(image_path):
     image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
     image = cv2.resize(image, (IMG_SIZE, IMG_SIZE))
     image = image / 255.0
+
     image = np.expand_dims(image, axis=0)
 
-    return image
+    return image.astype(np.float32)
